+++ src/main/java/com/airbnb/crud/airbnbDB/house/repository/HouseRRepository.java
@@ -7,5 +7,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @ReadOnlyRepository
-public interface HouseRRepository extends JpaRepository<House, Integer> {
+public interface HouseRRepository extends JpaRepository<House, Long> {
 }
