+++ src/main/java/com/airbnb/crud/airbnbDB/location/repository/LocationRWRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LocationRWRepository extends JpaRepository<Location, Integer> {
+public interface LocationRWRepository extends JpaRepository<Location, Long> {
 }
