+++ src/main/java/com/airbnb/crud/airbnbDB/person/repository/PersonRWRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PersonRWRepository extends JpaRepository<Person, Integer> {
+public interface PersonRWRepository extends JpaRepository<Person, Long> {
 }
